+++ src/components/pokemonCard/pokemonCard.module.css
@@ -1,11 +1,3 @@
-.root {
-    width: 268px;
-    height: 388px;
-    margin: 16px;
-    perspective: 1000px;
-    cursor: pointer;
-}
-
 .pokemonCard {
     --fire: #FDDFDF;
     --grass: #DEFDE0;
@@ -22,15 +14,14 @@
     --fighting: #E6E0D4;
     --normal:#F5F5F5;
 
-    width: inherit;
-    height: inherit;
+    width: 100%;
+    height: 100%;
     position: relative;
     transform-style: preserve-3d;
     transition: .6s .1s;
     transform: rotateY(180deg);
 }
 
-/* TODO: This place where we should change :hover on .active */
 .pokemonCard.active {
     animation: flipFront 1000ms 300ms forwards;
 }
@@ -42,6 +33,10 @@
     to {
         transform: rotateY(0deg);
     }
+}
+
+.pokemonCard.selected {
+    margin-top: -40px;
 }
 
 .cardFront,
@@ -93,6 +88,10 @@
 
 .wrap.back {
     z-index: -1;
+    background-image: url(../../assets/cardBack.jpg);
+    background-size: cover;
+    background-position: center;
+    background-repeat: no-repeat;
 }
 
 .cardBack {
