+++ src/routes/gamePage/routes/board/index.js
@@ -1,0 +1,2 @@
+import BoardPage from "./boardPage";
+export default BoardPage;
