+++ src/routes/gamePage/routes/finish/index.js
@@ -1,0 +1,9 @@
+const FinishPage = () => {
+    return (
+        <div>
+            This is finish page!
+        </div>
+    )
+}
+
+export default FinishPage;
