+++ src/routes/gamePage/routes/start/index.js
@@ -1,0 +1,2 @@
+import StartPage from "./startPage";
+export default StartPage;
